debug

+++ mmaction/datasets/pipelines/formating.py
@@ -17,7 +17,7 @@ def to_tensor(data):
     if isinstance(data, torch.Tensor):
         return data
     if isinstance(data, np.ndarray):
-        return torch.from_numpy(data).float()
+        return torch.from_numpy(data)
     if isinstance(data, Sequence) and not mmcv.is_str(data):
         return torch.tensor(data)
     if isinstance(data, int):
